split verifier index conversion to separate file
