+++ crypto/bindings/conversion-core.ts
@@ -20,6 +20,7 @@ import {
 
 export {
   ConversionCore,
+  ConversionCores,
   conversionCore,
   conversionCorePerField,
   freeOnFinalize,
@@ -45,6 +46,7 @@ type WasmClasses = {
 };
 
 type ConversionCore = ReturnType<typeof conversionCorePerField>;
+type ConversionCores = ReturnType<typeof conversionCore>;
 
 function conversionCore(wasm: wasm) {
   const fp = conversionCorePerField(wasm, {
